feat skus

+++ src/services/company.service.ts
@@ -12,47 +12,49 @@ CompanyService.prototype.getCompanies = async function() {
         const result = await this.companyRepository.getCompanies()
         return result.rows
     } catch(e) {
-        console.error(e.message)
-        throw new Error("Falha ao buscar empresas")
+        throw new Error("Falha ao buscar empresas: "+ e.message)
     }
 }
 
 CompanyService.prototype.getCompany = async function(companyId: number) {
     try {
         const result = await this.companyRepository.getCompany(companyId)
-        return result.rows
+        return result.rows[0]
     } catch(e) {
-        console.error(e.message)
-        throw new Error("Falha ao buscar empresa")
+        throw new Error("Falha ao buscar empresa: "+ e.message)
     }
 }
 
 CompanyService.prototype.saveCompany = async function(company: Company) {
     try {        
         const result = await this.companyRepository.saveCompany(company)
-        return result.rows
+        return result.rows[0]
     } catch(e) {
-        console.error(e.message)
-        throw new Error("Falha ao inserir empresa")
+        throw new Error("Falha ao inserir empresa: "+ e.message)
     }
 }
 
 CompanyService.prototype.updateCompany = async function(company: Company) {
     try {        
+        const comp = await this.getCompany(company.id)
+        if(!comp)
+            throw new Error("empresa nao encontrada")
         const result = await this.companyRepository.updateCompany(company)
-        return result.rows
+        return result.rows[0]
     } catch(e) {
-        console.error(e.message)
-        throw new Error("Falha ao atualizar empresa")
+        throw new Error("Falha ao atualizar empresa: "+ e.message)
     }
 }
 
 CompanyService.prototype.deleteCompany = async function(company: Company) {
     try {        
-        const result = await this.companyRepository.deleteCompany(company)
-        return result.rows
+        const comp = await this.getCompany(company.id)
+        if(!comp)
+            throw new Error("empresa nao encontrada")
+        await this.companyRepository.deleteCompany(company)
+        return {message: "empresa deleted"}
     } catch(e) {
-        throw new Error("Falha ao deletar empresa "+ e.message)
+        throw new Error("Falha ao deletar empresa: "+ e.message)
     }
 }
